Commit 3: Se añade funcionalidad a los botones. Ya se pueden grabar datos en la DB, tambien se leen las tablas al arranque del programa. Problemas detectados: Las tablas se borran cada vez que se arranca el programa. La tabla que muestra las tiendas no muestra las inserciones de datos. Pendiente de realizar: Añadir borrado de datos de las tablas. Crear las relaciones tiendas-productos y tiendas-empleados (las clases estan montadas, pero me han dado problemas hasta ahora).

+++ src/main/java/com/mycompany/adtarea04/modelos/TableProdXTiendaModel.java
@@ -104,7 +104,7 @@ public class TableProdXTiendaModel implements TableModel {
         Producto p = this.prodTienda.get(rowIndex);
         switch (columnIndex) {
             case 0:
-                p.setIdentificador((Integer) aValue);
+                p.setIdentificador((Long) aValue);
                 break;
             case 1:
                 p.setDescripcion((String) aValue);
